Module Workflow__.Types

type kv = Yaml.value

A Yaml key-value store -- this is exposed in order to be flexible with things like environment variable names which cannot be dynamically decided by the user if bound to a record type

val kv_to_yaml : kv -> Yaml.value

val kv_of_yaml : Yaml.value -> kv Yaml.res

type env = kv

The type of environment variables

type output = kv

The type of outputs

type run = {

run_shell : string;

run_workdir : string;

}

The type of run statements

type defaults = {

default_run : run;

}

The type of defaults

type container = {

image : string option;

credentials : kv option;

container_env : env option;

ports : int list option;

volumes : string list option;

}

The type of docker containers

type services = container

type with_ = kv

The type of input variables for external actions

type step = {

step_name : string option;

uses : string option;

step_env : env option;

step_workdir : string option;

step_shell : string option;

with_ : with_ option;

step_if : string option;

continue_on_err : bool option;

container : container option;

services : services option;

step_run : string option;

}

The type of steps in a job

type strategy = {

matrix : kv option;

fail_fast : bool option;

max_parallel : int option;

}

The type of strategies to define things like matrices

type job = {

job_name : string option;

strategy : strategy option;

runs_on : string;

outputs : output option;

job_env : env option;

job_defaults : defaults option;

job_if : string option;

steps : step list;

timeout : int option;

needs : string option;

}

The type of jobs

val job_to_yaml : job -> Yaml.value

val job_of_yaml : Yaml.value -> job Yaml.res

type on =

| List of string list

| Complex of Workflow.Events.t
The type of triggers for the workflow

type 'a t = {

name : string option;

on : on;

env : env option;

defaults : defaults option;

jobs : 'a;

}

The type of Github Workflows -- the parameter 'a allows you supply names to your jobs

val to_yaml : ('a -> Yaml.value) -> 'a t -> Yaml.value

val of_yaml : (Yaml.value -> 'a Yaml.res) -> Yaml.value -> 'a t Yaml.res