Module Workflow__Pp

val job : Stdlib.Format.formatter -> Workflow.Types.job -> unit

Pretty prints a job

val workflow : ?⁠drop_null:bool -> ('a -> Yaml.value) -> Stdlib.Format.formatter -> 'a Workflow.Types.t -> unit

Pretty prints a workflow. The user must supply a function for converting the type parameter to a yaml value. The drop_nulls argument can specifiy whether or not you want print null values in your yaml, the default being no