Module Workflow

module Types : sig ... end

Describes the workflow syntax as OCaml types

module Events : sig ... end

Describes the triggers for workflows

module Pp : sig ... end

Pretty printers

module Yaml_util : sig ... end

Yaml combinators

Default values and combinators

val simple_event : string list -> Types.on

Construct simple triggers for the workflow as a list such as [push, pull_request]

val complex_event : Events.t -> Types.on

Construct a complex trigger for workflows using events

Run

val run : Types.run

Default run shell (bash) and working directory (scripts)

val with_run_shell : string -> Types.run -> Types.run

val with_run_workdir : string -> Types.run -> Types.run

Containers & Services

val container : Types.container

Default empty container

val with_image : string -> Types.container -> Types.container

val with_credentials : Types.kv -> Types.container -> Types.container

val with_container_env : Types.kv -> Types.container -> Types.container

val with_ports : int list -> Types.container -> Types.container

val with_volumes : string list -> Types.container -> Types.container

Steps

val step : Types.step

Default empty step

val with_step_name : string -> Types.step -> Types.step

val with_uses : string -> Types.step -> Types.step

val with_step_env : Types.env -> Types.step -> Types.step

val with_step_workdir : string -> Types.step -> Types.step

val with_step_shell : string -> Types.step -> Types.step

val with_with : Types.with_ -> Types.step -> Types.step

val with_step_if : string -> Types.step -> Types.step

val with_continue_on_err : bool -> Types.step -> Types.step

val with_container : Types.container -> Types.step -> Types.step

val with_services : Types.services -> Types.step -> Types.step

val with_step_run : string -> Types.step -> Types.step

Strategy

val strategy : Types.strategy

Default empty strategy

val with_matrix : Types.kv -> Types.strategy -> Types.strategy

val with_fail_fast : bool -> Types.strategy -> Types.strategy

val with_max_parallel : int -> Types.strategy -> Types.strategy

Job

val job : string -> Types.job

job runs_on constructs a new job to run on runs_on -- note to get a list of items you should use the matrix approach with expr

val expr : string -> string

expr s constructs a Github Yaml expressions i.e. "${{ s }}"

val with_job_name : string -> Types.job -> Types.job

val with_strategy : Types.strategy -> Types.job -> Types.job

val with_runs_on : string -> Types.job -> Types.job

val with_outputs : Types.output -> Types.job -> Types.job

val with_job_env : Types.env -> Types.job -> Types.job

val with_job_defaults : Types.defaults -> Types.job -> Types.job

val with_job_if : string -> Types.job -> Types.job

val with_steps : Types.step list -> Types.job -> Types.job

val with_timeout : int -> Types.job -> Types.job

val with_needs : string -> Types.job -> Types.job

Workflow

val t : 'a -> 'a Types.t

t a makes a new workflow

val with_name : string -> 'a Types.t -> 'a Types.t

val with_on : Types.on -> 'a Types.t -> 'a Types.t

val with_env : Types.env -> 'a Types.t -> 'a Types.t

val with_defaults : Types.defaults -> 'a Types.t -> 'a Types.t

val with_jobs : 'a -> 'a Types.t -> 'a Types.t